docs(rfc0036): "whole file" not "whole hour"; u128 ratio print (review)
Three review fixes: the before/after narrative in the RFC status note
and §9.27 said the window materialises the "whole hour", but the
measurement is about materialising the whole *file* (the single-row-group
unsorted ingest parquet, no pruning) — reworded for precision, matching
the 1/1-groups table. And the win-ratio print casts to u128 so
before_bytes * 100 can't overflow if the fixture ever grows.

+++ crates/ourios-querier/tests/it/rfc0036_window_materialization.rs
@@ -524,7 +524,8 @@ async fn rfc0036_2_materialization_before_after() {
 
     // Integer ×100 ratio (avoids a float cast under clippy::pedantic); read
     // as `N.NNx`, e.g. 198 ⇒ before materialises 1.98× the after bytes.
-    let win_x100 = before_bytes * 100 / after_bytes.max(1);
+    // `u128` so the ×100 can't overflow if the fixture ever grows.
+    let win_x100 = u128::from(before_bytes) * 100 / u128::from(after_bytes.max(1));
     eprintln!(
         "RFC0036.2 before/after: uncompacted survivors {}/{} groups, {} bytes \
          (query scanned {}); compacted survivors {}/{} groups, {} bytes (query \
